requirements.txt
torch==2.0.1
torchvision==0.15.2
opencv-python==4.7.0.72
numpy==1.24.4
onnxruntime-gpu==1.16.0
insightface==0.7.3
pillow==9.5.0
customtkinter==5.2.0
pandas==2.0.2
openpyxl==3.1.2
scikit-learn==1.3.0
tqdm==4.65.0
matplotlib==3.7.1